annotation-tool/requirements.txt
Pillow==10.4.0
numpy==1.26.4
gradio==4.44.0
opencv-python==4.10.0.84
watchdog==5.0.2